+++ app/src/main/java/mustafaozhan/github/com/mycurrencies/base/BaseViewModel.kt
@@ -4,10 +4,10 @@ import androidx.lifecycle.ViewModel
 import io.reactivex.Completable
 import io.reactivex.Observable
 import io.reactivex.disposables.CompositeDisposable
+import mustafaozhan.github.com.mycurrencies.data.DataManager
 import mustafaozhan.github.com.mycurrencies.extensions.applySchedulers
 import mustafaozhan.github.com.mycurrencies.model.Currencies
 import mustafaozhan.github.com.mycurrencies.model.MainData
-import mustafaozhan.github.com.mycurrencies.repository.old.DataManager
 import org.joda.time.Duration
 import org.joda.time.Instant
