+++ app/src/main/java/mustafaozhan/github/com/mycurrencies/data/api/ApiService.kt
@@ -1,4 +1,4 @@
-package mustafaozhan.github.com.mycurrencies.base.api.backend
+package mustafaozhan.github.com.mycurrencies.data.api
 
 import io.reactivex.Observable
 import mustafaozhan.github.com.mycurrencies.model.Currencies
