+++ app/src/main/java/mustafaozhan/github/com/mycurrencies/data/preferences/GeneralSharedPreferences.kt
@@ -1,4 +1,4 @@
-package mustafaozhan.github.com.mycurrencies.repository.data
+package mustafaozhan.github.com.mycurrencies.data.preferences
 
 import com.google.gson.Gson
 import mustafaozhan.github.com.mycurrencies.base.BaseSharedPreferences
@@ -7,8 +7,11 @@ import mustafaozhan.github.com.mycurrencies.model.MainData
 import javax.inject.Inject
 import javax.inject.Singleton
 
+/**
+ * Created by Mustafa Ozhan on 7/10/18 at 9:50 PM on Arch Linux wit Love <3.
+ */
 @Singleton
-class DataRepository @Inject
+class GeneralSharedPreferences @Inject
 constructor() : BaseSharedPreferences() {
 
     companion object {
