+++ app/src/main/java/mustafaozhan/github/com/mycurrencies/di/ViewModelModule.kt
@@ -2,10 +2,10 @@ package mustafaozhan.github.com.mycurrencies.di
 
 import dagger.Module
 import dagger.Provides
+import mustafaozhan.github.com.mycurrencies.data.DataManager
 import mustafaozhan.github.com.mycurrencies.main.activity.MainActivityViewModel
 import mustafaozhan.github.com.mycurrencies.main.fragment.main.MainFragmentViewModel
 import mustafaozhan.github.com.mycurrencies.main.fragment.settings.SettingsFragmentViewModel
-import mustafaozhan.github.com.mycurrencies.repository.old.DataManager
 import mustafaozhan.github.com.mycurrencies.room.dao.CurrencyDao
 import mustafaozhan.github.com.mycurrencies.room.dao.OfflineRatesDao
 import mustafaozhan.github.com.mycurrencies.slider.SliderActivityViewModel
