--- app/src/main/java/mustafaozhan/github/com/mycurrencies/repository/api/ApiRepository.kt
@@ -1,3 +1,0 @@
-package mustafaozhan.github.com.mycurrencies.repository.api
-
-class ApiRepository
