+++ app/src/main/java/mustafaozhan/github/com/mycurrencies/slider/SliderActivityViewModel.kt
@@ -2,7 +2,7 @@ package mustafaozhan.github.com.mycurrencies.slider
 
 import io.reactivex.Completable
 import mustafaozhan.github.com.mycurrencies.base.BaseViewModel
-import mustafaozhan.github.com.mycurrencies.repository.old.DataManager
+import mustafaozhan.github.com.mycurrencies.data.DataManager
 
 class SliderActivityViewModel(
     override val dataManager: DataManager
